requirements.txt
streamlit
pandas
streamlit_option_menu
pdfminer.six 
nltk 
gensim 
sumy 
transformers 
bert-extractive-summarizer==0.5.0
wordcloud 
matplotlib 
pillow 
python-docx 
rouge_score 
bert-score 